Master rel 2019.1

+++ lib_parsing/DPB.h
@@ -1,6 +1,6 @@
 /******************************************************************************
 *
-* Copyright (C) 2018 Allegro DVT2.  All rights reserved.
+* Copyright (C) 2019 Allegro DVT2.  All rights reserved.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -129,16 +129,16 @@ typedef struct t_DpbNode
   uint8_t uNextDecOrder; /*!< Index of the   next   node in Decoding order  */
 
   /* info on the reference picture */
-  int iFramePOC; /*!< POC of this reference node */
+  int32_t iFramePOC; /*!< POC of this reference node */
   uint32_t slice_pic_order_cnt_lsb;
   AL_EMarkingRef eMarking_flag; /*!< status of this reference node */
 
-  int16_t iFrame_num;
-  int16_t iSlice_frame_num;
-  int16_t iFrame_num_wrap;
-  int16_t iLong_term_frame_idx;
-  int16_t iLong_term_pic_num;
-  int16_t iPic_num;
+  int32_t iFrame_num;
+  int32_t iSlice_frame_num;
+  int32_t iFrame_num_wrap;
+  int32_t iLong_term_frame_idx;
+  int32_t iLong_term_pic_num;
+  int32_t iPic_num;
 
   uint8_t pic_output_flag; /*!< whether picture must be displayed or not */
   bool bIsReset; /*!< Node has been reseted or not */
@@ -199,7 +199,7 @@ typedef struct t_DPB
   uint8_t uCurRef;
   uint8_t uCountRef;            /*!< Number of used node in the reference list */
   uint8_t uCountPic;            /*!< Number of used node in the reference list */
-  int16_t MaxLongTermFrameIdx; // used in picture marking process
+  int32_t MaxLongTermFrameIdx;  /*!< Number of the max long term index used in picture marking process */
   bool bLastHasMMCO5;
   AL_EDpbMode eMode; /*!< Possible DPB mode */
